+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/limit/LimitTest.java
@@ -80,17 +80,17 @@ public class LimitTest extends AbstractEntityViewTest {
     }
 
     @Test
-//    @Category({ NoH2.class, NoMySQLOld.class, NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class })
-    // H2 and the MySQL before 8 didn't support lateral joins which are required here
+//    @Category({ NoMySQLOld.class, NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class })
+    // MySQL before 8 didn't support lateral joins which are required here
     // EclipseLink, Datanucleus and OpenJPA don't support extended SQL
     // NOTE: Hibernate 6.3 bug: https://hibernate.atlassian.net/browse/HHH-17386
-    @Category({ NoH2.class, NoMySQLOld.class, NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class, NoHibernate60.class })
+    @Category({ NoMySQLOld.class, NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class, NoHibernate60.class })
     public void testLimitMultiset() {
         test(PersonLimitMultisetView.class, null);
     }
 
     @Test
-    @Category({ NoMySQLOld.class, NoHibernate42.class, NoHibernate43.class, NoHibernate50.class, NoHibernate51.class, NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class })
+    @Category({ NoHibernate42.class, NoHibernate43.class, NoHibernate50.class, NoHibernate51.class, NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class })
     // We need a left entity join for this so Hibernate < 5.1 can't be used
     // MySQL before 8 didn't support lateral and also don't support correlated LIMIT subqueries in quantified predicates
     // EclipseLink doesn't support subqueries in functions which is required for LIMIT
@@ -101,14 +101,14 @@ public class LimitTest extends AbstractEntityViewTest {
     }
 
     @Test
-//    @Category({ NoMySQLOld.class, NoHibernate42.class, NoHibernate43.class, NoHibernate50.class, NoHibernate51.class, NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class })
+//    @Category({ NoHibernate42.class, NoHibernate43.class, NoHibernate50.class, NoHibernate51.class, NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class })
     // We need a left entity join for this so Hibernate < 5.1 can't be used
     // MySQL before 8 didn't support lateral and also don't support correlated LIMIT subqueries in quantified predicates
     // EclipseLink doesn't support subqueries in functions which is required for LIMIT
     // Datanucleus fails because of a NPE?
     // OpenJPA has no function support
     // NOTE: Hibernate 6.3 bug: https://hibernate.atlassian.net/browse/HHH-17386
-    @Category({ NoMySQLOld.class, NoHibernate42.class, NoHibernate43.class, NoHibernate50.class, NoHibernate51.class, NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class, NoHibernate60.class })
+    @Category({ NoHibernate42.class, NoHibernate43.class, NoHibernate50.class, NoHibernate51.class, NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class, NoHibernate60.class })
     public void testLimitJoinExpression() {
         test(PersonLimitJoinExpressionView.class, null);
     }
@@ -123,7 +123,7 @@ public class LimitTest extends AbstractEntityViewTest {
     }
 
     @Test
-    @Category({ NoMySQLOld.class, NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class })
+    @Category({ NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class })
     // MySQL before 8 didn't support lateral and also don't support correlated LIMIT subqueries in quantified predicates
     // EclipseLink doesn't support subqueries in functions which is required for LIMIT
     // Datanucleus fails because of a NPE?
@@ -133,7 +133,7 @@ public class LimitTest extends AbstractEntityViewTest {
     }
 
     @Test
-    @Category({ NoMySQLOld.class, NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class })
+    @Category({ NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class })
     // MySQL before 8 didn't support lateral and also don't support correlated LIMIT subqueries in quantified predicates
     // EclipseLink doesn't support subqueries in functions which is required for LIMIT
     // Datanucleus fails because of a NPE?